pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clarify-rl"
version = "0.1.0"
description = "ClarifyRL: Train LLMs to ask clarifying questions instead of hallucinating — an OpenEnv RL environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "huggingface_hub>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
py-modules = ["models", "client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
import_mode = "importlib"
markers = [
    "integration: requires full openenv stack (deselect with -m 'not integration')",
]
